Split windows tests into different class, use ClassFixture to reduce setup cost
