+++ source/Tests/ShellExecutorFixture.cs
@@ -8,7 +8,6 @@ using FluentAssertions;
 using Octopus.Shellfish;
 using Octopus.Shellfish.Plumbing;
 using Octopus.Shellfish.Windows;
-using Tests.Plumbing;
 using Xunit;
 
 namespace Tests
@@ -18,7 +17,6 @@ namespace Tests
         // ReSharper disable InconsistentNaming
         const int SIG_TERM = 143;
         const int SIG_KILL = 137;
-        const string Username = "test-shellexecutor";
 
         static readonly string Command = PlatformDetection.IsRunningOnWindows ? "cmd.exe" : "bash";
         static readonly string CommandParam = PlatformDetection.IsRunningOnWindows ? "/c" : "-c";
@@ -106,125 +104,6 @@ namespace Tests
 
             exitCode.Should().Be(0, "the process should have run to completion");
             infoMessages.ToString().Should().ContainEquivalentOf("customvalue", "the environment variable should have been copied to the child process");
-            errorMessages.ToString().Should().BeEmpty("no messages should be written to stderr");
-        }
-
-        [WindowsFact]
-        public void DebugLogging_ShouldContainDiagnosticsInfo_DifferentUser()
-        {
-            var user = new TestUserPrincipal(Username);
-
-            var arguments = $"{CommandParam} \"echo %userdomain%\\%username%\"";
-            // Target the CommonApplicationData folder since this is a place the particular user can get to
-            var workingDirectory = Environment.GetFolderPath(Environment.SpecialFolder.CommonApplicationData);
-            var networkCredential = user.GetCredential();
-            var customEnvironmentVariables = new Dictionary<string, string>();
-
-            var exitCode = Execute(Command,
-                arguments,
-                workingDirectory,
-                out var debugMessages,
-                out var infoMessages,
-                out var errorMessages,
-                networkCredential,
-                customEnvironmentVariables,
-                CancellationToken);
-
-            exitCode.Should().Be(0, "the process should have run to completion");
-            debugMessages.ToString()
-                .Should()
-                .ContainEquivalentOf(Command, "the command should be logged")
-                .And.ContainEquivalentOf($@"{user.DomainName}\{user.UserName}", "the custom user details should be logged")
-                .And.ContainEquivalentOf(Environment.GetFolderPath(Environment.SpecialFolder.CommonApplicationData), "the working directory should be logged");
-            infoMessages.ToString().Should().ContainEquivalentOf($@"{user.DomainName}\{user.UserName}");
-            errorMessages.ToString().Should().BeEmpty("no messages should be written to stderr");
-        }
-
-        [WindowsFact]
-        public void RunningAsDifferentUser_ShouldCopySpecialEnvironmentVariables()
-        {
-            var user = new TestUserPrincipal(Username);
-
-            var arguments = $"{CommandParam} \"echo {EchoEnvironmentVariable("customenvironmentvariable")}\"";
-            // Target the CommonApplicationData folder since this is a place the particular user can get to
-            var workingDirectory = Environment.GetFolderPath(Environment.SpecialFolder.CommonApplicationData);
-            var networkCredential = user.GetCredential();
-            var customEnvironmentVariables = new Dictionary<string, string>
-            {
-                { "customenvironmentvariable", "customvalue" }
-            };
-
-            var exitCode = Execute(Command,
-                arguments,
-                workingDirectory,
-                out _,
-                out var infoMessages,
-                out var errorMessages,
-                networkCredential,
-                customEnvironmentVariables,
-                CancellationToken);
-
-            exitCode.Should().Be(0, "the process should have run to completion");
-            infoMessages.ToString().Should().ContainEquivalentOf("customvalue", "the environment variable should have been copied to the child process");
-            errorMessages.ToString().Should().BeEmpty("no messages should be written to stderr");
-        }
-
-        [WindowsFact]
-        public void RunningAsDifferentUser_ShouldWorkLotsOfTimes()
-        {
-            var user = new TestUserPrincipal(Username);
-
-            using var cts = new CancellationTokenSource(TimeSpan.FromSeconds(240));
-
-            for (var i = 0; i < 20; i++)
-            {
-                var arguments = $"{CommandParam} \"echo {EchoEnvironmentVariable("customenvironmentvariable")}%\"";
-                // Target the CommonApplicationData folder since this is a place the particular user can get to
-                var workingDirectory = Environment.GetFolderPath(Environment.SpecialFolder.CommonApplicationData);
-                var networkCredential = user.GetCredential();
-                var customEnvironmentVariables = new Dictionary<string, string>
-                {
-                    { "customenvironmentvariable", $"customvalue-{i}" }
-                };
-
-                var exitCode = Execute(Command,
-                    arguments,
-                    workingDirectory,
-                    out _,
-                    out var infoMessages,
-                    out var errorMessages,
-                    networkCredential,
-                    customEnvironmentVariables,
-                    cts.Token);
-
-                exitCode.Should().Be(0, "the process should have run to completion");
-                infoMessages.ToString().Should().ContainEquivalentOf($"customvalue-{i}", "the environment variable should have been copied to the child process");
-                errorMessages.ToString().Should().BeEmpty("no messages should be written to stderr");
-            }
-        }
-
-        [WindowsFact]
-        public void RunningAsDifferentUser_CanWriteToItsOwnTempPath()
-        {
-            var user = new TestUserPrincipal(Username);
-
-            var arguments = $"{CommandParam} \"echo hello > %temp%hello.txt\"";
-            // Target the CommonApplicationData folder since this is a place the particular user can get to
-            var workingDirectory = Environment.GetFolderPath(Environment.SpecialFolder.CommonApplicationData);
-            var networkCredential = user.GetCredential();
-            var customEnvironmentVariables = new Dictionary<string, string>();
-
-            var exitCode = Execute(Command,
-                arguments,
-                workingDirectory,
-                out _,
-                out _,
-                out var errorMessages,
-                networkCredential,
-                customEnvironmentVariables,
-                CancellationToken);
-
-            exitCode.Should().Be(0, "the process should have run to completion after writing to the temp folder for the other user");
             errorMessages.ToString().Should().BeEmpty("no messages should be written to stderr");
         }
 
@@ -334,60 +213,10 @@ namespace Tests
             infoMessages.ToString().Should().ContainEquivalentOf($@"{Environment.UserName}");
         }
 
-        [WindowsTheory]
-        [InlineData("powershell.exe", "-command \"Write-Host $env:userdomain\\$env:username\"")]
-        public void RunAsCurrentUser_PowerShell_ShouldWork(string command, string arguments)
-        {
-            var workingDirectory = "";
-            var networkCredential = default(NetworkCredential);
-            var customEnvironmentVariables = new Dictionary<string, string>();
-
-            var exitCode = Execute(command,
-                arguments,
-                workingDirectory,
-                out _,
-                out var infoMessages,
-                out var errorMessages,
-                networkCredential,
-                customEnvironmentVariables,
-                CancellationToken);
-
-            exitCode.Should().Be(0, "the process should have run to completion");
-            errorMessages.ToString().Should().BeEmpty("no messages should be written to stderr");
-            infoMessages.ToString().Should().ContainEquivalentOf($@"{Environment.UserDomainName}\{Environment.UserName}");
-        }
-
-        [WindowsTheory]
-        [InlineData("cmd.exe", "/c \"echo %userdomain%\\%username%\"")]
-        [InlineData("powershell.exe", "-command \"Write-Host $env:userdomain\\$env:username\"")]
-        public void RunAsDifferentUser_ShouldWork(string command, string arguments)
-        {
-            var user = new TestUserPrincipal(Username);
-
-            // Target the CommonApplicationData folder since this is a place the particular user can get to
-            var workingDirectory = Environment.GetFolderPath(Environment.SpecialFolder.CommonApplicationData);
-            var networkCredential = user.GetCredential();
-            var customEnvironmentVariables = new Dictionary<string, string>();
-
-            var exitCode = Execute(command,
-                arguments,
-                workingDirectory,
-                out _,
-                out var infoMessages,
-                out var errorMessages,
-                networkCredential,
-                customEnvironmentVariables,
-                CancellationToken);
-
-            exitCode.Should().Be(0, "the process should have run to completion");
-            errorMessages.ToString().Should().BeEmpty("no messages should be written to stderr");
-            infoMessages.ToString().Should().ContainEquivalentOf($@"{user.DomainName}\{user.UserName}");
-        }
-
         static string EchoEnvironmentVariable(string varName)
             => PlatformDetection.IsRunningOnWindows ? $"%{varName}%" : $"${varName}";
 
-        static int Execute(
+        public static int Execute(
             string command,
             string arguments,
             string workingDirectory,
